reorganize files and packages
