+++ .run/LinVAM.run.xml
@@ -12,7 +12,7 @@
     <option name="IS_MODULE_SDK" value="true" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
-    <option name="SCRIPT_NAME" value="src/linvam/linvam.py" />
+    <option name="SCRIPT_NAME" value="linvam/main.py" />
     <option name="PARAMETERS" value="--debug" />
     <option name="SHOW_COMMAND_LINE" value="false" />
     <option name="EMULATE_TERMINAL" value="false" />
