+++ linvam/mouse/_winmouse.py
