+++ linvam/pauseactioneditwnd.py
@@ -1,6 +1,6 @@
 from PyQt6.QtWidgets import QDialog
 
-from src.linvam.ui_pauseactioneditwnd import Ui_PauseActionEditDialog
+from linvam.ui_pauseactioneditwnd import Ui_PauseActionEditDialog
 
 
 class PauseActionEditWnd(QDialog):
