+++ linvam/ui/pauseactioneditwnd.ui
