+++ linvam/ui_commandeditwnd.py
