+++ linvam/util.py
